parallel controller (run-int-tests)
